+++ plugin-locationlayer/src/main/java/com/mapbox/mapboxsdk/plugins/locationlayer/LocationLayerConstants.java
@@ -22,15 +22,15 @@ class LocationLayerConstants {
   static final String LOCATION_SOURCE = "mapbox-location-source";
 
   // Layers
-  static final String LOCATION_LAYER = "mapbox-location-layer";
-  static final String LOCATION_BACKGROUND_LAYER = "mapbox-location-stroke-layer";
-  static final String LOCATION_ACCURACY_LAYER = "mapbox-location-accuracy-layer";
-  static final String LOCATION_BEARING_LAYER = "mapbox-location-bearing-layer";
-  static final String LOCATION_NAVIGATION_LAYER = "mapbox-location-navigation-layer";
+  static final String FOREGROUND_LAYER = "mapbox-location-layer";
+  static final String BACKGROUND_LAYER = "mapbox-location-stroke-layer";
+  static final String ACCURACY_LAYER = "mapbox-location-accuracy-layer";
+  static final String BEARING_LAYER = "mapbox-location-bearing-layer";
+  static final String NAVIGATION_LAYER = "mapbox-location-navigation-layer";
 
   // Icons
-  static final String USER_LOCATION_ICON = "mapbox-location-icon";
-  static final String USER_LOCATION_BEARING_ICON = "mapbox-location-bearing-icon";
-  static final String USER_LOCATION_BACKGROUND_ICON = "mapbox-location-stroke-icon";
-  static final String USER_LOCATION_PUCK_ICON = "mapbox-location-puck-icon";
+  static final String LOCATION_ICON = "mapbox-location-icon";
+  static final String BEARING_ICON = "mapbox-location-bearing-icon";
+  static final String BACKGROUND_ICON = "mapbox-location-stroke-icon";
+  static final String PUCK_ICON = "mapbox-location-puck-icon";
 }