+++ plugin-places/src/main/java/com/mapbox/plugins/places/autocomplete/data/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Contains classes directly related to the SQLite database for search history.
+ *
+ * @since 0.1.0
+ */
+package com.mapbox.plugins.places.autocomplete.data;